+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- only visible POIs are shown
 - can show GPX logs
 - can process whole photo gallery in non blocking way (>6k photos)
 - not blocking gallery processing